backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
pandas
openpyxl
python-multipart
python-jose[cryptography]
passlib[bcrypt]
